new version system; major braces bug fix; minor updates

+++ src/backend/ast/nodes.rs
@@ -43,7 +43,7 @@ impl fmt::Debug for BinaryOpNode {
 Number Node
 */
 pub struct NumberNode {
-    pub number: i32,
+    pub number: i64,
 }
 
 impl fmt::Debug for NumberNode {
